Move DefaultHostBuilderFactory to new AppBuilding namespace
