+++ src/AppMotor.CliApp/AppMotor.CliApp.NuGet.props
@@ -20,7 +20,8 @@
     <CommonlyUsedTypes Include="AppMotor.CliApp.CommandLine.CliCommand" />
     <CommonlyUsedTypes Include="AppMotor.CliApp.CommandLine.CliParam" />
     <CommonlyUsedTypes Include="AppMotor.CliApp.CommandLine.Hosting.GenericHostCliCommand" />
-    <CommonlyUsedTypes Include="AppMotor.CliApp.CommandLine.Hosting.DefaultHostBuilderFactory" />
+    <CommonlyUsedTypes Include="AppMotor.CliApp.CommandLine.AppBuilding.DefaultHostBuilderFactory" />
+    <CommonlyUsedTypes Include="AppMotor.CliApp.CommandLine.CliApplication" />
 
     <CommonlyUsedTypes Include="AppMotor.CliApp.Terminals.Terminal" />
     <CommonlyUsedTypes Include="AppMotor.CliApp.Terminals.Formatting.TermText" />
