+++ tests/AppMotor.HttpServer.Tests/Tests/HttpsTests.cs
@@ -1,8 +1,8 @@
 ﻿// SPDX-License-Identifier: MIT
 // Copyright AppMotor Framework (https://github.com/skrysmanski/AppMotor)
 
+using AppMotor.CliApp.AppBuilding;
 using AppMotor.CliApp.CommandLine;
-using AppMotor.CliApp.CommandLine.Hosting;
 using AppMotor.CliApp.HttpServer.TestUtils;
 using AppMotor.Core.Certificates;
 using AppMotor.Core.Net;
